\name{SeriestiempoBMR-package}
\alias{SeriestiempoBMR-package}
\alias{SeriestiempoBMR}
\docType{package}
\title{
\packageTitle{SeriestiempoBMR}
}
\description{
\packageDescription{SeriestiempoBMR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SeriestiempoBMR}
\packageIndices{SeriestiempoBMR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SeriestiempoBMR}

Maintainer: \packageMaintainer{SeriestiempoBMR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
